requirements.txt
Pillow>=10.0.0,<11.0.0
reportlab>=4.0.0,<5.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=7.0.0,<9.0.0
